Refactor (rename)wq code
