+++ examples/BUILD
@@ -1,0 +1,11 @@
+cc_binary(
+  name = "example",
+  srcs = [
+    "main.cpp",
+    "main.h",
+  ],
+  deps = [
+    "//libs:LRUCache",
+    "//utils:print",
+  ],
+)
