+++ examples/main.h
@@ -1,7 +1,7 @@
 #ifndef MAIN_H
 #define MAIN_H
 
-#include "LRUCache/LRUCache.h"
+#include "libs/LRUCache.h"
 #include "utils/print.h"
 
 #endif
