--- src/BUILD
@@ -1,23 +1,0 @@
-# cc_library(
-#   name = "main",
-#   srcs = [
-#     "main.cpp",
-#     "main.h",
-#   ],
-#   deps = [
-#     "//LRUCache",
-#     "//utils:print",
-#   ],
-# )
-
-cc_binary(
-  name = "mainProg",
-  srcs = [
-    "main.cpp",
-    "main.h",
-  ],
-  deps = [
-    "//LRUCache",
-    "//utils:print",
-  ],
-)
